Add macro for auto-complete -> company source conversion
There is a macro under foreign-elisp/ac-company.el that converts an
existing company source to an auto-complete source.  I attempted the
opposite, a macro that converts an existing auto-complete source to a
company source.  Unfortunately it doesn't work yet for the auto-complete
source I'm most interested in, the one jedi.el provides, probably
because jedi.el does use auto-complete in non-standard ways, like using
the init function every time a completion is requested instead of an
one-shot setup.  Nonetheless basic usage for jedi.el is included in case
I figure out a workaround.