Partially implement CIDER-style clean up
CIDER has an interesting feature for dealing with REPL output that slows
it down considerably.  Instead of just clearing the entire buffer by
setting the maximum buffer size to zero and telling comint to truncate
it, it deletes the last output generated by the last command and
replaces it with an appropriate comment.  This is a much more sensible
thing to do since it is pretty likely that the last command executed
made the REPL slow down, not its entire contents.  A well executed
version could make it upstream and could be available for every
comint-derived mode.